Extract run display component
